Translation model/requirements.txt
tensorflow==2.11.1
numpy==1.24.2
uvicorn==0.21.1
pydantic==1.10.6
python-dotenv==1.0.0
python-multipart==0.0.6
fastapi==0.95.0
keras==2.11.0
Keras-Preprocessing==1.1.2
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
# tensorflow-estimator==2.11.0
# tensorflow-intel==2.11.1
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
